scripts/requirements.txt
# UniFlow AI 采集服务依赖

# 核心框架
Flask==3.1.2
flask-cors==6.0.2

# AI 服务
openai>=1.0.0

# 数据库
supabase>=2.0.0

# 网页抓取
beautifulsoup4==4.14.3
requests==2.32.5
playwright>=1.40.0  # 可选，用于微信公众号链接抓取

# 图片处理（OCR）
pytesseract==0.3.13
Pillow>=12.0.0

# 工具库
python-dotenv==1.2.1

# 数据处理
pandas>=2.0.0  # Excel 导入功能
